LibWeb: Move DeclarationOrAtRule code into DeclarationOrAtRule.cpp
